+++ dnsimple/zones_records_test.go
@@ -65,12 +65,12 @@ func TestZonesService_ListRecords_WithOptions(t *testing.T) {
 		httpResponse := httpResponseFixture(t, "/listZoneRecords/success.http")
 
 		testQuery(t, r, url.Values{
-			"page":        []string{"2"},
-			"per_page":    []string{"20"},
-			"sort":        []string{"name,expiration:desc"},
-			"name":        []string{"example"},
-			"name_like":   []string{"www"},
-			"record_type": []string{"A"},
+			"page":      []string{"2"},
+			"per_page":  []string{"20"},
+			"sort":      []string{"name,expiration:desc"},
+			"name":      []string{"example"},
+			"name_like": []string{"www"},
+			"type":      []string{"A"},
 		})
 
 		w.WriteHeader(httpResponse.StatusCode)