+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - NEW: Added support for the DNSSEC Beta (GH-58)
 
 - CHANGED: Changed response types to not be exported (GH-54)
+- CHANGED: Updated registrar URLs (GH-59)
 
 #### Release 0.14.0